Update The Auto Hunt Bot
After a certain number of Afk-ing sessions, I was confident that the bot could still detect you due to long-term auto-hunting.
So I added a little code that will pause the code for 10 minutes and keep doing so. Keep in mind that if you use it a lot, you may or may not get banned because there is no bypass. This bot is simple, and I believe all I did was trick the bot into thinking we were hunting the bot ourselves.